force dashboard renaming
