+++ src/components/dashy/GeneralInfo.svelte
