+++ src/components/dashy/StoryMetrics.svelte
