+++ src/components/dashy/dream/create/Scenario.svelte
