+++ src/components/dashy/dream/create/Slider.svelte
