+++ src/pages/dashboard/dream/create/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Layout.astro';
 
-import CreateStory from '@components/dashboard/dream/CreateStory.svelte';
+import CreateStory from '@components/dashy/dream/CreateStory.svelte';
 
 const arrow = true;
