+++ src/pages/dashboard/dream/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Layout.astro';
 
-import Dream from '@components/dashboard/Dream.svelte';
+import Dream from '@components/dashy/Dream.svelte';
 
 const arrow = true;
