+++ src/pages/dashboard/dream/manage/[topic]/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Layout.astro';
 
-import ExploreStory from '@components/dashboard/dream/manage/ExploreStory.svelte';
+import ExploreStory from '@components/dashy/dream/manage/ExploreStory.svelte';
 
 const { topic } = Astro.params;
