+++ src/pages/dashboard/dream/manage/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Layout.astro';
 
-import ManageStories from '@components/dashboard/dream/ManageStories.svelte';
+import ManageStories from '@components/dashy/dream/ManageStories.svelte';
 
 const { topic } = Astro.params;
