+++ src/pages/dashboard/profile/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from '@layouts/Layout.astro';
 
-import Profile from '@components/dashboard/Profile.svelte';
+import Profile from '@components/dashy/Profile.svelte';
 
 const arrow = true;
